Start moving BTES functionality to separate submodule Geothermal
