+++ src/facility/wells/wells.jl
@@ -405,7 +405,6 @@ end
 
 include("mswells.jl")
 include("stdwells.jl")
-include("btes.jl")
 
 include("stdwells_equations.jl")
